Add remaining functions and a pulsar grid struct.
This adds the remainder of the necessary functions to generate data.
Some of them are currently coded with static constant parameters, which
should change in the future, but other than that everything is mostly
done.

The pulsar grid datastructure basically creates a data structure, which
holds the spherical polar coordinates of the pulsar as viewed from the
Solar System Baricentre (SSB). We calculate the cartesians on request,
which might not be more efficient, but the interface is better and we do
not have to store a lot of data. This might change in the future if I
see that the performance depends on this a lot.

Signed-off-by: Ignas Anikevicius (gns_ank) <anikevicius@gmail.com>