Finish writing data generation functions
The Python code is written, but there are many basic Cython
optimizations left to be done. Also I need to write some Python "glue"
code, which we would use to manipulate the data.

TODO: write unit tests to check the functions I write, as otherwise
there can be lots of bugs.

Signed-off-by: Ignas Anikevicius (gns_ank) <anikevicius@gmail.com>