Add a separate schedule generation function
Here I simplify some of the functions, so that there are less of them.
Mainly, I can now generate a schedule and read it at a later time. This
is a first step towards a working Reduced Basis construction algorithm.

Also, I have added one parameter to a function, so that I can turn of
noise generation. This is very useful for signal template generation.

Signed-off-by: Ignas Anikevicius (gns_ank) <anikevicius@gmail.com>