RB construction: Hopefully everything is done
Hopefully the algorithm is more or less complete. However, no testing
was done and some enhancements could be done as well because at the
moment the traversal through the parameter space is very stupid and
expensive. Therefore, there should be a better way.

I need to understand the hierarchical nature of the RB method
construction.

Signed-off-by: Ignas Anikevicius (gns_ank) <anikevicius@gmail.com>