Lots of bugfixing and speed improvements
I wanted to make lots of speed improvements and I failed due to some
Cython errors, which annoy me. I will quickly (probably in one day or
so) recode everything in C++.

My experiment with Cython, I guess, is semi-failed (or
semi-accomplished), as it is really good for prototyping, but when you
need really fine grained control, it falls short.

Signed-off-by: Ignas Anikevicius (gns_ank) <anikevicius@gmail.com>