Add a new function to evaluate the residual norm quickly.
This will use less computational power than the previous algorithms as
we are not doing any expensive inner product calculatinos. It is only
a couple of dot products. Also this version tries to use CBLAS to speed
up the dot product evaluations.

Update the greedy algorithm to use the new optimized function as well.
We are now calculating the norms of the trial signal templates from the
start, so that we do not have to do it later. Also, we can output them
into a file, so that when we are doing Bayesian analysis, we do not have
to recalculate them.

Signed-off-by: Ignas Anikevicius (gns_ank) <anikevicius@gmail.com>